maybe fixes a bug with reddit api?
I didn't tried it yet, but according to the documentation, the promise returned by `getNewModmailConversation(id)` gives us a `ModmailConversation` object, so we don't need to call `.fetch()`.

You probably went wrongly with `getNewModmailConversations()` which gives a `Listing` object, hence the need to call `.fetch()` on it.

See: https://not-an-aardvark.github.io/snoowrap/snoowrap.html#getNewModmailConversation__anchor