refactors checkIfProcessed()
The dockblock says this function returns true or false, but that's a lie! (as cakes, beware of cakes...)
Since we query the database, the result will necessarily be asyncronous (a promise in our case, because of the `.then()` after the query).

I rewritten your code in order to handle that promise, and make your function actually return a boolean value.

For a promise primer, see: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise