refactors checkIfEvent()
This function should also return a boolean value according to its name and dockblock.

A best practice is to have one purpose for each function: checkIfEvent should just check if the given post is effectively an event or not, but it shouldn't run db queries or stuff like that (otherwise, it'll be too complicated to grasp).

By rewriting your code, we see that the `main()` function starts to be "cluttered". It means it's time to refactor it a bit more!